+++ test/index.spec.js
@@ -7,10 +7,7 @@ let test = require('blue-tape')
 let http = require('http')
 let Buffer = require('safe-buffer').Buffer
 
-let linereader = require('../lib/linereader')
-let lambda = require('../lib/lambda')
-let kms = require('../lib/kms')
-let cerberus = require('../index')
+let CerberusClient = require('../index')
 
 let testPort = process.env.TEST_PORT || 3032
 let cerberusHost = 'http://localhost:' + testPort
@@ -33,15 +30,6 @@ let defaultToken = {
   'client_token': Math.floor(Math.random() * (1e6 + 1)),
   'lease_duration': 1,
   'renewable': 'true'
-}
-let lambdaContext = {
-  invokedFunctionArn: 'arn:aws:lambda:us-east-1:1234567890:function:NU-cerberus-test'
-}
-let lambdaConfiguration = {
-  'FunctionName': 'NU-cerberus-test',
-  'FunctionArn': 'arn:aws:lambda:us-east-1:123456789:function:NU-cerberus-test',
-  'Role': 'arn:aws:iam::123456789:role/lambda_basic_execution',
-  'Description': 'Test lambda for NU-cerberus'
 }
 
 let mockCerberusHost = (action, handlerOrValue) => {
@@ -75,33 +63,13 @@ let mockCerberusHost = (action, handlerOrValue) => {
   })
 }
 
-test('Apps put', spec => {
+test('Cerberus Client', spec => {
   let sandbox
-  let lineReaderUser = null
-  let lineReaderPassword = null
-  let lineReaderMfa = null
-  let lambdaResult = null
-  let lambdaStub = null
-  let ksmResult = null
   let setup = () => {
-    lambdaResult = lambdaConfiguration
-    ksmResult = defaultToken
     sandbox = sinon.sandbox.create()
-    sandbox.stub(linereader, 'readLine')
-      .onCall(0).resolves(lineReaderUser)
-      .onCall(1).resolves(lineReaderPassword)
-      .onCall(2).resolves(lineReaderMfa)
-      .rejects('too many calls to linereader')
-    lambdaStub = sandbox.stub(lambda, 'getFunctionConfiguration').resolves(lambdaResult)
-    sandbox.stub(kms, 'decrypt').resolves(ksmResult)
   }
   let teardown = () => {
     sandbox.restore()
-    lineReaderUser = null
-    lineReaderPassword = null
-    lineReaderMfa = null
-    lambdaResult = null
-    ksmResult = null
   }
   let teardownError = (err) => {
     teardown()
@@ -109,13 +77,13 @@ test('Apps put', spec => {
   }
 
   spec.test('module loads', t => {
-    t.ok(cerberus, 'module loaded')
+    t.ok(CerberusClient, 'module loaded')
     t.end()
   })
 
   spec.test('verifies options', t => {
-    t.throws(() => cerberus(), /options parameter/, 'requires options')
-    t.throws(() => cerberus({ aws: {} }), /options.hostUrl/, 'requires hostUrl')
+    t.throws(() => new CerberusClient(), /options parameter/, 'requires options')
+    t.throws(() => new CerberusClient({ aws: {} }), /options.hostUrl/, 'requires hostUrl')
     t.end()
   })
 
@@ -123,129 +91,24 @@ test('Apps put', spec => {
     let original = process.env.CERBERUS_TOKEN
     let testToken = 'test token'
     process.env.CERBERUS_TOKEN = testToken
-    let client = cerberus({ hostUrl: cerberusHost, debug: false })
+    let client = new CerberusClient({ hostUrl: cerberusHost, debug: false })
     setup()
 
-    return mockCerberusHost(() => client.get('test'))
+    return mockCerberusHost(() => client.getSecureData('test'))
       .then(() => {
-        t.equal(mockCalls[0].req.headers['x-vault-token'], testToken, 'Vault header uses CERBERUS_TOKEN')
+        t.equal(mockCalls[0].req.headers['x-cerberus-token'], testToken, 'Cerberus Auth Token header uses CERBERUS_TOKEN')
         process.env.CERBERUS_TOKEN = original
-      })
-      .then(teardown, teardownError)
-  })
-
-  spec.test('uses environment variable for hostUrl when present', t => {
-    let original = process.env.CERBERUS_ADDR
-    process.env.CERBERUS_ADDR = cerberusHost
-    let client = cerberus({ lambdaContext, debug: false })
-
-    setup()
-    t.plan(1)
-
-    return mockCerberusHost(() => client.get('test'))
-      .then(result => {
-        process.env.CERBERUS_ADDR = original
-        t.ok(mockCalls.length, 'http called')
-      })
-      .then(teardown, teardownError)
-  })
-
-  spec.test('uses lambdaContext from constructor', t => {
-    let client = cerberus({ lambdaContext, hostUrl: cerberusHost })
-    setup()
-    t.plan(2)
-
-    return mockCerberusHost(() => client.get('test'))
-      .then(result => {
-        t.ok(lambdaStub.called, 'lambda was called')
-        t.ok(mockCalls.length, 'http called')
-      })
-      .then(teardown, teardownError)
-  })
-
-  spec.test('uses lambdaContext from setLambdaContext', t => {
-    let client = cerberus({ hostUrl: cerberusHost })
-    client.setLambdaContext(lambdaContext)
-
-    setup()
-    t.plan(2)
-
-    return mockCerberusHost(() => client.get('test'))
-      .then(result => {
-        t.ok(lambdaStub.called, 'lambda was called')
-        t.ok(mockCalls.length, 'http called')
-      })
-      .catch(err => {
-        t.comment(err)
-        t.fail('error from cerberus')
-      })
-      .then(teardown, teardownError)
-  })
-
-  spec.test('Uses ECS metadata endpoint when detected', (t) => {
-    process.env.AWS_CONTAINER_CREDENTIALS_RELATIVE_URI = 'foo/bar'
-
-    let client = cerberus({ hostUrl: cerberusHost, debug: true, ecsMetadataUrl: `${cerberusHost}/v2/iam/metadata` })
-    setup()
-    t.plan(1)
-
-    let metadataCalled = false
-
-    const handler = (req, res) => {
-      if (req.url.indexOf('v2/iam/metadata') !== -1) {
-        metadataCalled = true
-        res.end(JSON.stringify({
-          RoleArn: 'arn:aws:iam::123456789:role/BrowseServiceRole',
-          AccessKeyId: 'abc123',
-          SecretAccessKey: 'def456',
-          SessionToken: 't28038'
-        }))
-      } else {
-        res.end(JSON.stringify(defaultCerberusResponse))
-      }
-    }
-    return mockCerberusHost(() => client.get('test'), handler)
-    .then(() => {
-      t.true(metadataCalled, 'metadata endpoint not called')
-    })
-    .then(teardown, teardownError)
-  })
-
-  spec.test('Prompt flow prompts if config option is set and other methods fail', t => {
-    process.env.CERBERUS_TOKEN = undefined
-    t.plan(2) // the 2nd assertion is present to ensure the http server closes before the test completes
-
-    let client = cerberus({ hostUrl: cerberusHost, lambdaContext, prompt: true, debug: false })
-
-    lineReaderUser = 'user'
-    lineReaderPassword = 'password'
-    setup()
-
-    const action = () => client.get('test')
-    const handler = (req, res) => {
-      res.setHeader('Content-Type', 'application/json')
-      if (req.url.indexOf('auth/user') !== -1) {
-        t.equal('Basic ' + Buffer.from('user:password').toString('base64'), req.headers.authorization, 'sent prompted credentials')
-        res.end(JSON.stringify({ data: { client_token: defaultToken } }))
-      } else {
-        res.end(JSON.stringify(defaultCerberusResponse))
-      }
-    }
-
-    return mockCerberusHost(action, handler)
-      .then(result => {
-        t.pass('letting the promise finish so the server can close properly')
       })
       .then(teardown, teardownError)
   })
 
   spec.test('If request returns an errors array treat it like an error in the post-getToken flow', t => {
     process.env.CERBERUS_TOKEN = defaultToken
-    let client = cerberus({ hostUrl: cerberusHost, aws: {} })
+    let client = new CerberusClient({ hostUrl: cerberusHost })
 
     setup()
 
-    const action = () => client.put('some/test/path', 'System expects a JSON object, not a string, this should fail')
+    const action = () => client.writeSecureData('some/test/path', 'System expects a JSON object, not a string, this should fail')
     const handler = (req, res) => {
       res.setHeader('Content-Type', 'application/json')
       res.end(JSON.stringify({ errors: [ 'Failed to parse JSON input: json: cannot unmarshal string into Go value of type map[string]interface {}' ] }))
@@ -264,11 +127,11 @@ test('Apps put', spec => {
   })
 
   spec.test('If cerberus returns error from auth response return error', t => {
-    let client = cerberus({ lambdaContext, hostUrl: cerberusHost })
+    let client = new CerberusClient({ hostUrl: cerberusHost })
     setup()
     t.plan(1)
 
-    return mockCerberusHost(() => client.get('test'), { error_id: 'f887ba2a-d104-4323-93e0-d22304932f56', errors: [ { code: 99216, message: 'The specified IAM role is not valid.' } ] })
+    return mockCerberusHost(() => client.getSecureData('test'), { error_id: 'f887ba2a-d104-4323-93e0-d22304932f56', errors: [ { code: 99216, message: 'The specified IAM role is not valid.' } ] })
       .then(result => {
         console.log('test result', result)
         t.fail()
@@ -282,11 +145,11 @@ test('Apps put', spec => {
 
   spec.test('If cerberus returns error from key response return error', t => {
     process.env.CERBERUS_TOKEN = 'testToken'
-    let client = cerberus({ lambdaContext, hostUrl: cerberusHost })
+    let client = new CerberusClient({ hostUrl: cerberusHost })
     setup()
     t.plan(1)
 
-    return mockCerberusHost(() => client.get('test'), { error_id: 'f887ba2a-d104-4323-93e0-d22304932f56', errors: [ { code: 3243, message: 'Key request failure' } ] })
+    return mockCerberusHost(() => client.getSecureData('test'), { error_id: 'f887ba2a-d104-4323-93e0-d22304932f56', errors: [ { code: 3243, message: 'Key request failure' } ] })
       .then(result => {
         process.env.CERBERUS_TOKEN = undefined
         console.log('test result', result)
@@ -302,11 +165,11 @@ test('Apps put', spec => {
 
   spec.test('Get key should work', t => {
     process.env.CERBERUS_TOKEN = 'testToken'
-    let client = cerberus({ lambdaContext, hostUrl: cerberusHost })
+    let client = new CerberusClient({ hostUrl: cerberusHost })
 
     t.plan(1)
 
-    return mockCerberusHost(() => client.get('test'), { data: { success: 'someKey' } })
+    return mockCerberusHost(() => client.getSecureData('test'), { data: { success: 'someKey' } })
       .then(result => {
         process.env.CERBERUS_TOKEN = undefined
         t.same(result, { success: 'someKey' }, 'key returned')
@@ -322,7 +185,7 @@ test('Apps put', spec => {
 
   spec.test('Get file should work', t => {
     process.env.CERBERUS_TOKEN = 'testToken'
-    let client = cerberus({ lambdaContext, hostUrl: cerberusHost })
+    let client = new CerberusClient({ hostUrl: cerberusHost })
 
     t.plan(1)
 
@@ -341,11 +204,11 @@ test('Apps put', spec => {
   spec.test('should send package version', t => {
     process.env.CERBERUS_TOKEN = 'testToken'
     let packageData = require('../package.json')
-    let client = cerberus({ lambdaContext, hostUrl: cerberusHost })
+    let client = new CerberusClient({ hostUrl: cerberusHost })
 
     t.plan(1)
 
-    return mockCerberusHost(() => client.get('test'), { data: { success: 'someKey' } })
+    return mockCerberusHost(() => client.getSecureData('test'), { data: { success: 'someKey' } })
       .then(result => {
         process.env.CERBERUS_TOKEN = undefined
         t.equal(mockCalls[0].req.headers['x-cerberus-client'], `CerberusNodeClient/${packageData.version}`, 'version is sent to cerbeurs')
@@ -360,7 +223,7 @@ test('Apps put', spec => {
   })
 
   spec.test('If cerberus returns 404 response return error', t => {
-    let client = cerberus({ lambdaContext, hostUrl: cerberusHost })
+    let client = new CerberusClient({ hostUrl: cerberusHost })
     setup()
     t.plan(1)
 
@@ -377,7 +240,7 @@ test('Apps put', spec => {
       res.end(JSON.stringify(result))
     }
 
-    return mockCerberusHost(() => client.get('test'), handler)
+    return mockCerberusHost(() => client.getSecureData('test'), handler)
       .then(result => {
         console.log('test result', result)
         t.fail()
@@ -389,11 +252,11 @@ test('Apps put', spec => {
   })
 
   spec.test('If cerberus returns empty response return error', t => {
-    let client = cerberus({ lambdaContext, hostUrl: cerberusHost })
+    let client = new CerberusClient({ hostUrl: cerberusHost })
     setup()
     t.plan(1)
 
-    return mockCerberusHost(() => client.get('test'), {})
+    return mockCerberusHost(() => client.getSecureData('test'), {})
       .then(result => {
         console.log('test result', result)
         t.fail()
@@ -406,7 +269,7 @@ test('Apps put', spec => {
   })
 
   spec.test('If cerberus returns empty response return error using callbacks', t => {
-    let client = cerberus({ lambdaContext, hostUrl: cerberusHost })
+    let client = new CerberusClient({ hostUrl: cerberusHost })
     setup()
 
     t.plan(1)
@@ -420,7 +283,7 @@ test('Apps put', spec => {
     })
 
     server.listen(testPort, () => {
-      client.get('test', (err, result) => {
+      client.getSecureData('test', (err, result) => {
         server.close()
         t.comment(err)
         t.ok(/cannot decrypt token/.test(err && err.message), 'error from auth result')